TeacherPlan table PersonPrivilege table

+++ src/main/java/com/peter/kist/repository/TeacherPlanRepository.java
@@ -3,5 +3,5 @@ package com.peter.kist.repository;
 import com.peter.kist.model.entity.TeacherPlan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TeacherPlanRepository extends JpaRepository<TeacherPlan, Long> {
+public interface TeacherPlanRepository extends JpaRepository<TeacherPlan, Integer> {
 }
